# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: "Know Your Customer — BIAN Semantic API"
  description: >
    BIAN Service Domain 'Know Your Customer' (Process) on the 'KYC Assessment
    Procedure' control record. Screening pipeline: watchlist → REJECTED;
    missing documents / high-risk jurisdiction → REFERRED (analyst decides);
    otherwise APPROVED. Verdicts are published and, when a callbackUrl is
    given, delivered to the requesting account SD's kyc-result endpoint.
  version: 0.2.0
servers:
  - url: /
  - url: /sd-know-your-customer
    description: through the platform gateway
paths:
...
  /v1/kyc-assessment-procedure/{assessmentId}/control:
    put:
      operationId: decideReferredCase
      summary: "Control — analyst decision on a REFERRED case (reason mandatory; audit)"
      parameters: [{ $ref: "#/components/parameters/AssessmentId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, reason]
              properties:
                action: { type: string, enum: [approve, reject] }
                reason: { type: string }
      responses:
        "200": { description: Decided assessment (manuallyDecided=true) }
        "400": { $ref: "#/components/responses/Invalid" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/RuleViolation" }
  /v1/kyc-assessment-procedure/watchlist:
    get:
      operationId: getWatchlist
      summary: Current watchlist
      responses:
        "200":
          description: Watchlisted customer references
          content:
            application/json:
              schema:
                type: array
                items: { type: string }
    post:
      operationId: addToWatchlist
      summary: Add a customer reference to the watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customerReference]
              properties:
                customerReference: { type: string }
      responses:
        "201": { description: Added }
        "400": { $ref: "#/components/responses/Invalid" }
components:
  parameters:
    AssessmentId:
      name: assessmentId
      in: path
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Unknown assessment
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RuleViolation:
      description: Manual decision on a non-REFERRED case
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Invalid:
      description: Bad input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        code: { type: string, example: NOT_REFERRED }
        message: { type: string }